zynqmp-smmu: reduce size of sid cnode
The zynqmp utilizes 15 bits for its smmu stream ids but does not
allocate these values in a contiguous order resulting in a lot of unused
memory in a cnode array allocated by the maximum sid value. These
changes compress the amount of memory used for the zynqmp sid cnode
array by mapping cnode slots to the master id instead of the full sid
value.